apps/backend-python/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0
geoalchemy2==0.17.1

# Geoespacial
geopandas==0.14.1
shapely==2.0.2
folium==0.15.0
geopy==2.4.1
rasterio==1.3.9
pyproj==3.6.1

# Analytics & ML
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# OCR & Vision - Versiones compatibles
pytesseract==0.3.10
opencv-python==4.10.0.84
Pillow==10.1.0
pdf2image==1.16.3

# Weather & External APIs
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Authentication
clerk-backend-api==0.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0